Homolog

+++ app/app/Http/Controllers/Admin/ProductController.php
@@ -16,6 +16,10 @@ class ProductController extends Controller
     public function __construct(ProductRepositoryInterface $productRepository)
     {
         $this->repository = $productRepository;
+        $this->middleware("can:products.show", ['only' => ['index','show','search']]);
+        $this->middleware("can:products.create", ['only' => ['create','store']]);
+        $this->middleware("can:products.edit", ['only' => ['edit','update']]);
+        $this->middleware("can:products.destroy", ['only' => ['destroy']]);
     }
 
     public function index()
@@ -68,7 +72,7 @@ class ProductController extends Controller
     public function destroy($id) 
     {
         $product = $this->repository->getById($id);
-        if(Storage::exists($product->image)) {
+        if(!empty($product->image) && Storage::exists($product->image)) {
             Storage::delete($product->image);
         }
         $this->repository->delete($id);
